update egg drop puzzle

+++ src/DynamicProgram/EggDroppingPuzzle.hpp
@@ -5,7 +5,7 @@
 /// </summary>
 /// <param name="num_of_eggs">number of eggs.</param>
 /// <param name="num_of_floors">number of floors</param>
-/// <returns></returns>
+/// <returns>minimum trials with given eggs.</returns>
 size_t
 EggDropping(size_t num_of_eggs, size_t num_of_floors)
 {
@@ -27,7 +27,7 @@ EggDropping(size_t num_of_eggs, size_t num_of_floors)
 	{
 		for (j = 2; j <= num_of_floors; j++)
 		{
-			trials[i][j] = INT_MAX;
+			trials[i][j] = SIZE_MAX;
 			for (x = 1; x <= j; x++)
 			{
 				res = 1 + std::max(trials[i - 1][x - 1],
